# docs/swagger.yaml
basePath: /api
definitions:
...
  request_models.AddFeedbackRequest:
    properties:
      comment:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    required:
    - comment
    - rating
    - user_id
    type: object
...
  request_models.CreatePaymentRequest:
    properties:
      plan_code:
        type: string
    required:
    - plan_code
    type: object
...
  request_models.ForgotPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - email
    - new_password
    - token
    type: object
  request_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  request_models.RequestForgotPassword:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  request_models.RequestVerifyOtpToken:
    properties:
      email:
        type: string
      token:
        type: string
    required:
    - email
    - token
    type: object
  request_models.SignUpRequest:
    properties:
      display_name:
        maxLength: 50
        minLength: 3
        type: string
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - display_name
    - email
    - password
    type: object
...
  response_models.FeedbackResponse:
    properties:
      comment:
        type: string
      created_at:
        type: integer
      id:
        type: string
      rating:
        type: integer
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
      trace_id:
        type: string
    type: object
host: api.vivu-travel.site
info:
  contact: {}
  description: This is the API documentation for Vivu Travel Platform
  title: Vivu Travel API
  version: "1.0"
paths:
  /accounts/all:
    get:
      consumes:
      - application/json
      description: Fetch a list of all user accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all accounts
      tags:
      - Accounts
  /accounts/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the provided email if it exists
      parameters:
      - description: Forgot password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.RequestForgotPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Request a password reset
      tags:
      - Accounts
  /accounts/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Login to an account
      tags:
      - Accounts
  /accounts/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Account registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Register a new account
      tags:
      - Accounts
  /accounts/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using a valid OTP token
      parameters:
      - description: Password reset payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Reset password with OTP
      tags:
      - Accounts
  /accounts/verify-otp:
    post:
      consumes:
      - application/json
      description: Validates the provided OTP token for account verification
      parameters:
      - description: OTP token verification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.RequestVerifyOtpToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Verify an OTP token
      tags:
      - Accounts
...
  /feedback/add:
    post:
      consumes:
      - application/json
      description: Add a comment and rating for the app
      parameters:
      - description: Feedback payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.AddFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Add feedback
      tags:
      - Feedback
  /feedback/list:
    get:
      description: Get a paginated list of feedback
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response_models.FeedbackResponse'
            type: array
      summary: List feedback
      tags:
      - Feedback
...
  /payments/create-checkout:
    post:
      consumes:
      - application/json
      description: Create a checkout request for a subscription plan
      parameters:
      - description: Create Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_models.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a checkout request for a subscription plan
      tags:
      - Payments
...
  /payments/subscription-details:
    get:
      consumes:
      - application/json
      description: Retrieve subscription details for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get subscription details for the authenticated user
      tags:
      - Payments
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"